split into core, read and write
