+++ opcua_plugin/core_connect.go
@@ -64,7 +64,7 @@ func parseBase64PEMBundle(b64PemBundle string) (*tls.Certificate, error) {
 // - To dynamically generate client options that match the server’s security requirements.
 // - To handle different authentication methods, such as anonymous or username/password-based logins.
 // - To generate and include client certificates when using enhanced security policies like Basic256Sha256/Basic256/Basic128Rsa15
-func (g *OPCUAInput) GetOPCUAClientOptions(
+func (g *OPCUAConnection) GetOPCUAClientOptions(
 	selectedEndpoint *ua.EndpointDescription,
 	selectedAuthentication ua.UserTokenType,
 	discoveryOnly bool,
@@ -182,7 +182,7 @@ func (g *OPCUAInput) GetOPCUAClientOptions(
 // This function provides comprehensive logging of an endpoint’s properties, including security
 // configurations and server details. It is useful for debugging and auditing purposes, allowing
 // developers and administrators to verify endpoint settings and ensure they meet the required standards.
-func (g *OPCUAInput) LogEndpoint(endpoint *ua.EndpointDescription) {
+func (g *OPCUAConnection) LogEndpoint(endpoint *ua.EndpointDescription) {
 	g.Log.Infof("  EndpointURL: %s", endpoint.EndpointURL)
 	g.Log.Infof("  SecurityMode: %v", endpoint.SecurityMode)
 	g.Log.Infof("  SecurityPolicyURI: %s", endpoint.SecurityPolicyURI)
@@ -224,7 +224,7 @@ func (g *OPCUAInput) LogEndpoint(endpoint *ua.EndpointDescription) {
 }
 
 // LogEndpoints logs information for a slice of OPC UA endpoints.
-func (g *OPCUAInput) LogEndpoints(endpoints []*ua.EndpointDescription) {
+func (g *OPCUAConnection) LogEndpoints(endpoints []*ua.EndpointDescription) {
 	g.Log.Infof("Fetched %d endpoints", len(endpoints))
 	for i, endpoint := range endpoints {
 		g.Log.Infof("Endpoint %d:", i+1)
@@ -237,7 +237,7 @@ func (g *OPCUAInput) LogEndpoints(endpoints []*ua.EndpointDescription) {
 // not "None". If so it returns the first endpoint. The first endpoint will
 // always be the most "secure", because we order the endpoints by their security-
 // policy and -mode.
-func (g *OPCUAInput) checkForSecurityEndpoints(
+func (g *OPCUAConnection) checkForSecurityEndpoints(
 	ctx context.Context,
 	endpoints []*ua.EndpointDescription,
 	authType ua.UserTokenType,
@@ -289,7 +289,7 @@ func (g *OPCUAInput) checkForSecurityEndpoints(
 //     using the most suitable endpoint.
 //   - To handle servers that may return limited or DNS-specific endpoint information by substituting
 //     with user-specified host details.
-func (g *OPCUAInput) FetchAllEndpoints(ctx context.Context) ([]*ua.EndpointDescription, error) {
+func (g *OPCUAConnection) FetchAllEndpoints(ctx context.Context) ([]*ua.EndpointDescription, error) {
 	g.Log.Infof("Querying OPC UA server at: %s", g.Endpoint)
 
 	endpoints, err := opcua.GetEndpoints(ctx, g.Endpoint)
@@ -328,7 +328,7 @@ func (g *OPCUAInput) FetchAllEndpoints(ctx context.Context) ([]*ua.EndpointDescr
 //   - To ensure the client connects to the correct server(when encrypted)
 //   - To inform the user using a Fingerprint of the Server's certificate would
 //     be more secure
-func (g *OPCUAInput) checkServerCertificateFingerprint(endpoint *ua.EndpointDescription) error {
+func (g *OPCUAConnection) checkServerCertificateFingerprint(endpoint *ua.EndpointDescription) error {
 
 	if g.ServerCertificates[endpoint] == "" {
 		return fmt.Errorf("The servers endpoint '%s' doesn't provide any server certificate. "+
@@ -371,7 +371,7 @@ func (g *OPCUAInput) checkServerCertificateFingerprint(endpoint *ua.EndpointDesc
 // **Why this Function is needed:**
 //   - to store each endpoint of the OPC-UA-Server and it's corresponding fingerprint
 //   - to prevent edge-cases where certificates could differ for its endpoints
-func (g *OPCUAInput) storeServerCertificateFingerprint(endpoint *ua.EndpointDescription) error {
+func (g *OPCUAConnection) storeServerCertificateFingerprint(endpoint *ua.EndpointDescription) error {
 
 	pemCert := pem.EncodeToMemory(&pem.Block{
 		Type:  "CERTIFICATE",
@@ -409,7 +409,7 @@ func (g *OPCUAInput) storeServerCertificateFingerprint(endpoint *ua.EndpointDesc
 // **Why This Function is Needed:**
 // - To expand the list of available endpoints beyond the initial discovery when only one endpoint is found.
 // - To ensure comprehensive endpoint discovery, especially for servers that may not provide extensive endpoint lists.
-func (g *OPCUAInput) handleSingleEndpointDiscovery(ctx context.Context, endpoint *ua.EndpointDescription) ([]*ua.EndpointDescription, error) {
+func (g *OPCUAConnection) handleSingleEndpointDiscovery(ctx context.Context, endpoint *ua.EndpointDescription) ([]*ua.EndpointDescription, error) {
 	if endpoint == nil || endpoint.Server == nil || len(endpoint.Server.DiscoveryURLs) == 0 {
 		if endpoint != nil && endpoint.Server != nil && len(endpoint.Server.DiscoveryURLs) == 0 { // This is the edge case when there is no discovery URL
 			g.Log.Infof("No discovery URL. This is the endpoint: %v", endpoint)
@@ -463,7 +463,7 @@ func (g *OPCUAInput) handleSingleEndpointDiscovery(ctx context.Context, endpoint
 // This function iterates through a list of OPC UA endpoints and replaces their existing hostnames with a
 // new host provided by the user. This is particularly useful for addressing connectivity issues caused
 // by DNS name discrepancies or when routing through specific network configurations.
-func (g *OPCUAInput) ReplaceHostInEndpoints(endpoints []*ua.EndpointDescription, newHost string) ([]*ua.EndpointDescription, error) {
+func (g *OPCUAConnection) ReplaceHostInEndpoints(endpoints []*ua.EndpointDescription, newHost string) ([]*ua.EndpointDescription, error) {
 	var updatedEndpoints []*ua.EndpointDescription
 
 	for _, endpoint := range endpoints {
@@ -485,7 +485,7 @@ func (g *OPCUAInput) ReplaceHostInEndpoints(endpoints []*ua.EndpointDescription,
 // This function parses the provided endpoint URL, substitutes the host component with a new host, and reconstructs
 // the URL while preserving the original path and query parameters. It ensures that the modified URL remains
 // valid and compatible with OPC UA communication protocols.
-func (g *OPCUAInput) ReplaceHostInEndpointURL(endpointURL, newHost string) (string, error) {
+func (g *OPCUAConnection) ReplaceHostInEndpointURL(endpointURL, newHost string) (string, error) {
 
 	// Remove the "opc.tcp://" prefix to simplify parsing.
 	newHost = strings.TrimPrefix(newHost, "opc.tcp://")
@@ -514,7 +514,7 @@ func (g *OPCUAInput) ReplaceHostInEndpointURL(endpointURL, newHost string) (stri
 // This function decodes and parses the server’s certificate from PEM format and logs various attributes,
 // such as validity periods, DNS names, IP addresses, and URIs. It provides transparency into the security
 // credentials presented by the OPC UA server, aiding in the verification of server authenticity and trustworthiness.
-func (g *OPCUAInput) logCertificateInfo(certBytes []byte) {
+func (g *OPCUAConnection) logCertificateInfo(certBytes []byte) {
 	g.Log.Infof("  Server certificate:")
 
 	// Decode the certificate from base64 to DER format
@@ -551,7 +551,7 @@ func (g *OPCUAInput) logCertificateInfo(certBytes []byte) {
 // Returns:
 // - *opcua.Client: A pointer to the connected OPC UA client, if successful.
 // - error: An error object if the connection attempt fails.
-func (g *OPCUAInput) connectWithoutSecurity(
+func (g *OPCUAConnection) connectWithoutSecurity(
 	ctx context.Context,
 	endpoints []*ua.EndpointDescription,
 	authType ua.UserTokenType,
@@ -611,7 +611,7 @@ func (g *OPCUAInput) connectWithoutSecurity(
 // Returns:
 // - *opcua.Client - The connected OPC UA client.
 // - error - An error if the connection fails.
-func (g *OPCUAInput) encryptedConnect(
+func (g *OPCUAConnection) encryptedConnect(
 	ctx context.Context,
 	endpoints []*ua.EndpointDescription,
 	authType ua.UserTokenType,
@@ -651,7 +651,7 @@ func (g *OPCUAInput) encryptedConnect(
 // unencryptedConnect is used as a fallback, so if the user decides not to use
 // an encrypted connection this will be executed and tries to connect to any
 // endpoints ("encrypted">none-encrypted)
-func (g *OPCUAInput) unencryptedConnect(
+func (g *OPCUAConnection) unencryptedConnect(
 	ctx context.Context,
 	endpoints []*ua.EndpointDescription,
 	authType ua.UserTokenType,
@@ -683,7 +683,7 @@ func (g *OPCUAInput) unencryptedConnect(
 // connectToDirectEndpoint establishes a connection to an OPC UA server using a direct endpoint.
 // It takes a context for managing the connection lifecycle and an authentication type for user token.
 // It returns an OPC UA client instance and an error if the connection fails.
-func (g *OPCUAInput) connectToDirectEndpoint(ctx context.Context, authType ua.UserTokenType) (*opcua.Client, error) {
+func (g *OPCUAConnection) connectToDirectEndpoint(ctx context.Context, authType ua.UserTokenType) (*opcua.Client, error) {
 
 	g.Log.Infof("Directly connecting to the endpoint %s", g.Endpoint)
 
@@ -740,7 +740,7 @@ func (g *OPCUAInput) connectToDirectEndpoint(ctx context.Context, authType ua.Us
 //
 // Returns:
 // - error: An error if the connection fails, otherwise nil.
-func (g *OPCUAInput) connect(ctx context.Context) error {
+func (g *OPCUAConnection) connect(ctx context.Context) error {
 	var (
 		c         *opcua.Client
 		endpoints []*ua.EndpointDescription
@@ -809,7 +809,7 @@ func (g *OPCUAInput) connect(ctx context.Context) error {
 //
 // On failure:
 //   - it returns the error
-func (g *OPCUAInput) openConnection(
+func (g *OPCUAConnection) openConnection(
 	ctx context.Context,
 	endpoint *ua.EndpointDescription,
 	authType ua.UserTokenType,
