+++ opcua_plugin/core_helper.go
