+++ opcua_plugin/core_node_attributes.go
