+++ opcua_plugin/write.go
