+++ opcua_plugin/write_test.go
